Moving Core class to its own module (#1938)
